Refactor: 라우팅 처리부 app.js에서 분리
